consolidate s3 testing tools into one file
